[OBSDATA-1786]: Adding Channel acquire time metric for netty HTTP client (#237)
