+++ server/src/test/java/org/apache/druid/server/initialization/BaseJettyTest.java
@@ -25,6 +25,8 @@ import com.google.inject.Key;
 import com.google.inject.servlet.GuiceFilter;
 import org.apache.druid.guice.annotations.Self;
 import org.apache.druid.java.util.common.lifecycle.Lifecycle;
+import org.apache.druid.java.util.emitter.core.NoopEmitter;
+import org.apache.druid.java.util.emitter.service.ServiceEmitter;
 import org.apache.druid.java.util.http.client.HttpClient;
 import org.apache.druid.java.util.http.client.HttpClientConfig;
 import org.apache.druid.java.util.http.client.HttpClientInit;
@@ -128,7 +130,8 @@ public abstract class BaseJettyTest
                             .withReadTimeout(Duration.ZERO)
                             .withEagerInitialization(true)
                             .build(),
-            druidLifecycle
+            druidLifecycle,
+            new ServiceEmitter("", "", new NoopEmitter())
         );
       }
       catch (Exception e) {
